Refactor test suite and add more tests.
